Improvements to ActiveRecord::Rollback docs
The main problem I wanted to solve is that the "Call tech support" message in the example is
confusing in this scenario, because it's addressing the user, but the user will never see it.
I'm not sure if I'm not understanding something, or maybe this was copy/pasted from
another example years ago.

Here I remove that and also expand the docs, hopefully illustrating the concept better.